Provide some support for empty Arrays and Dictionaries
Previously, dealing with an empty array or dictionary would just lead to
a panic, as the code would assume at least one element existed.  This
was needed so that we could infer the type signature for D-Bus.  We
still need to know the type signature, so this commit creates wrapper
struct's for Array and Dictionary.  These classes also keep track of the
type signature.  If an Array or Dictionary will always be non-empty,
then the new() constructor can be used and it will infer the type
signature.  Otherwise, you must use new_with_sig() and provide the
signature yourself.